requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
pathspec==0.7.0
psycopg2==2.8.4
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
regex==2019.12.20
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
wrapt==1.11.2
